Uploading Helper class + pkl with shaped features for model input
