+++ models/best_model.h5.dvc
@@ -1,5 +1,5 @@
 outs:
-- md5: 6bd4f4420cdc14049370e8c4be84c05d
-  size: 11576232
+- md5: 39d3d2e4293e8e312ac9fe6884a5cfb0
+  size: 18916264
   hash: md5
   path: best_model.h5
